体验增强的修改，增加图片预加载

+++ war/src/js/gallery/setupPicture.js
@@ -3,11 +3,12 @@
 /*global G:false, $:false, alert:false */
 
 //@import "./setupLayout.js";
+//@import "../lib/loadImages.js";
 
 G.def(
     './gallery/setupPicture',
-    ['./setupLayout'],
-    function(setupLayout) {
+    ['./setupLayout', 'loadImages'],
+    function(setupLayout, loadImages) {
         'use strict';
 
         // --- load pics ---
@@ -22,6 +23,11 @@ G.def(
             oldCursor,
             cursor,
             picsApi = {
+                /**
+                 * 获取图片
+                 * @param {string} c cursor
+                 * @param {function} cb callback
+                 */
                 getPicsApi: function(c, cb) {
                     $.ajax({
                         url: '/api/pic/',
@@ -29,30 +35,55 @@ G.def(
                         success: function(data) {
                             var html = [],
                                 sameCursor = cursor === data.cursor,
-                                noMoreData = (data.pics.length < pageSize) || sameCursor;
+                                noMoreData = (data.pics.length < pageSize) || sameCursor,
+                                urls = [];
 
                             if (!sameCursor) {
                                 G.each(data.pics, function(v) {
+                                    urls.push(v.url);
                                     html.push(G.format(picTmpl, v));
                                 });
+                                // preload image
+                                loadImages(urls, function() {
+                                    cb.call(this, html.join(''), noMoreData);
+                                    cursor = data.cursor;
+                                });
+                            } else {
+                                cb.call(this, '', noMoreData);
                             }
-
-                            cb.call(this, html.join(''), noMoreData);
-                            cursor = data.cursor;
                         },
                         cache: false,
                         dataType: 'json'
                     });
                 },
+                lock: false,
+                /**
+                 * 加载更多图片
+                 */
                 loadMorePics: function () {
+                    if (picsApi.lock) {
+                        return;
+                    }
+                    picsApi.lock = true;
+                    $pullUp.css('opacity', 0.5);
                     picsApi.getPicsApi(cursor, function(html, noMoreData) {
                         $pics.append(html);
                         if (noMoreData) {
                             $pullUp.hide();
                         }
+                        $pullUp.css('opacity', 1);
+                        picsApi.lock = false;
                     });
                 },
+                /**
+                 * 更新图片
+                 */
                 updatePics: function() {
+                    if (picsApi.lock) {
+                        return;
+                    }
+                    picsApi.lock = true;
+                    $pullUp.css('opacity', 0.5);
                     picsApi.getPicsApi(null, function(html, noMoreData) {
                         $pics.html(html);
                         if (noMoreData) {
@@ -60,8 +91,13 @@ G.def(
                         } else {
                             $pullUp.show();
                         }
+                        $pullUp.css('opacity', 1);
+                        picsApi.lock = false;
                     });
                 },
+                /**
+                 * 在最前面插入图片
+                 */
                 prependPicture: function(data) {
                     $pics.prepend(G.format(picTmpl, data));
                 }
@@ -75,6 +111,7 @@ G.def(
             $startEdit = $('#startEdit'),
             startEditing = false;
 
+        // 进入与离开编辑模式
         $startEdit.click(function() {
             if (startEditing) {
                 startEditing = false;
@@ -86,9 +123,13 @@ G.def(
                 $container.addClass('start-edit');
             }
         });
+        // 删除图片
         $container.delegate('a[data-do=deletePics]', 'click', function() {
-            if (startEditing) {
+            if (startEditing && !this.lock) {
+                this.lock = true;
                 var $this = $(this);
+                $this.parent()
+                    .css('opacity', 0.5);
                 $.ajax({
                     url: '/api/pic/?id=' + $this.data('id'),
                     type: 'DELETE',
@@ -99,7 +140,12 @@ G.def(
                                 .remove();
                         } else {
                             alert('服务器出错，请刷新后重试');
+                            this.lock = false;
                         }
+                    },
+                    error: function() {
+                        alert('删除图片出错，请刷新后重试');
+                        this.lock = false;
                     }
                 });
             }
